requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
sqlalchemy
pydantic
pydantic-settings
profanity-check
google-generativeai
python-dotenv
